Allow setting attribute datablock values directly.
Issue #114.

+++ src/mmscenegraph/rust/src/constant.rs
@@ -20,7 +20,7 @@
 
 use nalgebra;
 
-pub type AttrIndex = u64;
+pub type AttrIndex = usize;
 pub type HashValue = u64;
 pub type NodeIndex = usize;
 pub type FrameValue = u32;
